Add files via upload
# Dimensionality-Reduction-with-PCA-and-LDA
In this project, Principal Component Analysis and Linear Discriminant 
Analysis are discussed as dimensionality reduction techniques.
They are implemented from scratch on iris dataset.

Specifically, PCA1, PCA2 and PCA3 are extracted and new features are projected to 1D, 2D and 3D spaces with the help of these vectors
respectively. 
LDA is implemented with only the first component, and features are projected to 1D with the help of this vector.

Conclusion:
LDA is much more successful in linearly separating classes than PCA, as it is a supervised machine learning technique and
already knows which data point belongs to which class, whereas PCA is an unsupervised machine learning technique and does 
not have that knowledge. Still, both of them can project data points to new dimensions.